Feature/api gateway support

+++ tests/test_runtime.py
@@ -1,44 +1,51 @@
 import json
 import unittest
 
-from tests.aws_lambda import LocalLambdaServer, start_local_lambda
+import boto3
+
+from tests.sam import LocalLambdaServer, start_local_lambda
 
 
 class TestRuntimeLayer(unittest.TestCase):
     lambda_server: LocalLambdaServer = None
 
     @classmethod
+    def isLocal(cls) -> bool:
+        return True
+
+    @classmethod
     def setUpClass(cls):
-        cls.lambda_server = start_local_lambda()
+        if cls.isLocal():
+            cls.lambda_server = start_local_lambda()
 
     def test_script(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="ExampleFunction", Payload=json.dumps({'x': 1}))
         raw_payload = response['Payload'].read().decode('utf-8')
-        json_payload = json.loads(raw_payload)
-        result = json_payload['result']
+        result = json.loads(raw_payload)
         self.assertEqual(result, 2)
 
+    def get_client(self):
+        return self.lambda_server.get_client() if self.isLocal() else boto3.client('lambda')
+
     def test_lowercase_extension(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="LowerCaseExtensionFunction", Payload=json.dumps({'x': 1}))
         raw_payload = response['Payload'].read().decode('utf-8')
-        json_payload = json.loads(raw_payload)
-        result = json_payload['result']
+        result = json.loads(raw_payload)
         self.assertEqual(result, 2)
 
     def test_multiple_arguments(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         payload = {'x': 'bar', 'y': 1}
         response = lambda_client.invoke(FunctionName="MultipleArgumentsFunction", Payload=json.dumps(payload))
         raw_payload = response['Payload'].read().decode('utf-8')
-        json_payload = json.loads(raw_payload)
-        result = json_payload['result']
+        result = json.loads(raw_payload)
         self.assertDictEqual(result, payload)
 
     @unittest.skip('Lambda local does not pass errors properly')
     def test_missing_source_file(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="MissingSourceFileFunction", Payload=json.dumps({'y': 1}))
         raw_payload = response['Payload'].read().decode('utf-8')
         json_payload = json.loads(raw_payload)
@@ -47,7 +54,7 @@ class TestRuntimeLayer(unittest.TestCase):
 
     @unittest.skip('Lambda local does not pass errors properly')
     def test_missing_function(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="MissingFunctionFunction", Payload=json.dumps({'y': 1}))
         raw_payload = response['Payload'].read().decode('utf-8')
         json_payload = json.loads(raw_payload)
@@ -56,7 +63,7 @@ class TestRuntimeLayer(unittest.TestCase):
 
     @unittest.skip('Lambda local does not pass errors properly')
     def test_missing_argument(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="ExampleFunction")
         raw_payload = response['Payload'].read().decode('utf-8')
         json_payload = json.loads(raw_payload)
@@ -65,7 +72,7 @@ class TestRuntimeLayer(unittest.TestCase):
 
     @unittest.skip('Lambda local does not pass errors properly')
     def test_unused_argument(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="ExampleFunction", Payload=json.dumps({'x': 1, 'y': 1}))
         raw_payload = response['Payload'].read().decode('utf-8')
         json_payload = json.loads(raw_payload)
@@ -74,7 +81,7 @@ class TestRuntimeLayer(unittest.TestCase):
 
     @unittest.skip('Lambda local does not pass errors properly')
     def test_long_argument(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         payload = {x: x for x in range(0, 100000)}
         response = lambda_client.invoke(FunctionName="VariableArgumentsFunction", Payload=json.dumps(payload))
         raw_payload = response['Payload'].read().decode('utf-8')
@@ -83,7 +90,7 @@ class TestRuntimeLayer(unittest.TestCase):
 
     @unittest.skip('Lambda local does not pass errors properly')
     def test_missing_library(self):
-        lambda_client = self.lambda_server.get_client()
+        lambda_client = self.get_client()
         response = lambda_client.invoke(FunctionName="MissingLibraryFunction", Payload=json.dumps({'y': 1}))
         raw_payload = response['Payload'].read().decode('utf-8')
         json_payload = json.loads(raw_payload)
@@ -92,4 +99,5 @@ class TestRuntimeLayer(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
-        cls.lambda_server.kill()
+        if cls.isLocal():
+            cls.lambda_server.kill()
